Defaults and Rao-Blackwellization changes to ipmcmc
The Rao-Blackwellization has been changed when not using all the
samples so that the, effectively free, RB of the gibbs updates is
always performed even if only returning a single sample for each
sweep.

Changes to default options are also added so that the "out-of-the-box"
behaviour is more sensible.  In particular, all particles are now
returned by default and the default number of nodes has been set
to 32.